perf(git): set depth on fetches
We've seen that fetching can be bottleneck for large repos. By
restricting the depth of the fetches we can speedup this process.

For the target we can simply fetch with depth 1, because all we need is
a commit to branch from and cherry-pick commits onto.

For the other fetches (baseref and headref) we need enough depth to find
the common ancestor between the two. Generally most commits are probably
found on the baseref (depending on the number of commits merged into the
base, and the number of commits in the current PR to merge).

In the last week nixpkg had 891 commits merged to its default branch.
As long as PRs that need to be backported are merged within a week, this
would hold. But it might be a bit tight. We should probably make this
configurable. Or better still, not require this fetch.